Basic Keyboard Functionality
Added basic functionality for the keyboard.  It makes keys as buttons that send a corresponding char value when clicked.  Has support for special buttons, such as the included SHIFT key that can alter the other keys or perform other effects.  The keys resize when you click them, with frequently used keys becoming larger and less-used keys becoming smaller, with a maximum and minimum scaling factor of 2.0 and 0.5 respectively.

The json and Path imports are for later on, when I will be adding support for saving and loading sets of keys.