JSON save and load
The program now saves the current keyboard configuration when the user closes the window, so that it will be the same when they reopen the program later.  Scales of the keys are now saved along with the keyboard layout.  There is now a default_keyboard.json which the program will use to start a new keyboard if the saved_keyboard.json is deleted.  The program now uses __file__ to ensure that the .json files will be loaded even when the program is run in pipenv from the main folder.