Basic Loot Boxes
Loot boxes now unlock after completing a word.  The word detection works with any non-letter seperating the word.  (However currently, any whitespace causes an assert error, pls fix maku)
The unlocked keys are added to the keyboard, and any longer keys have extra space added after them.  Right now, it will add this space even if the key was at the end of the row and the space was not required.  This might be addressable in the future.  Also fixed enter's key_description entry which erroneously had it print the text "enter" instead of a "\n" character.