Lootbox Graphic & Popup
Lootbox unlock now has a custom popup window that displays all the unlocked keys at once.  It uses a 1-second delay to create a simple animation of the capsule opening.  Cleaned up some old code.  Changed the spacebar to only be size 2 instead of 14, as it was causing too much unneeded blank space.