Menu Buttons
New file, save, and load are now accessible via on-screen buttons as well as from the alt menu.  Also changed loading behavior so that it backspaces once after loading the file - this deletes the finaly \n added at the end of the file so that the cursor continues from where it left off instead of on a new line.